rtos.desc: try to install module on whatever kernel dir is installed
Do not get the kernel module version from the module, but from the
running armadillo instead.
Building a module for the kernel package works but the module version
embedded in the module might not be exactly the same (localversion),
so do not rely on it but get 'uname -r' instead.

This means the kernel should not be updated in the same swu