README: add note about remoteproc stop errata
AN5317 states as follow:

On i.MX 8M platforms, "remoteproc" stops only the Cortex-M CPU, not
the Cortex-M system.
Therefore, any in-flight Cortex-M bus transactions would hang after
the CPU is halted and this can only be resolved with a full SoC reset.
It is not recommended to stop the Cortex-M7 CPU in a production system.
If the system must stop the Cortex-M7 CPU, reload the image and restart
it. Then, the Cortex-M7 CPU must be in the WFI state and have no external
access to the Cortex-M7 TCM memory through eDMA or other similar
transactions. A possible solution is to implement a handshake between
the Cortex-M and Cortex-A CPUs to confirm that the Cortex-M CPU is safe
to stop or reset.

warn users remoteproc is not safe for production.

Link: https://www.nxp.com/docs/en/application-note/AN5317.pdf
Change-Id: I674acd5462bdf64c3765f5dd139abb6d430fa5e1